+++ Editor/ReleaseJson.cs
@@ -19,4 +19,41 @@ namespace Hananoki.GitHubDownload {
 		[NonSerialized]
 		public bool toggle;
 	}
+
+
+
+	[Serializable]
+	public class Tags {
+		public string name;
+		public string tarball_url;
+		public string zipball_url;
+		[Serializable]
+		public class Commit {
+			public string sha;
+			public string url;
+		}
+		public Commit commit;
+
+		public string node_id;
+	}
+
+	public static class TagsExtention {
+		public static Tags GetTags( this Tags[] tags, string hash ) {
+			foreach( var p in tags ) {
+				if( p.commit.sha == hash ) {
+					return p;
+				}
+			}
+			return null;
+		}
+
+		public static string GetRevisionHash( this Tags[] tags, string revision ) {
+			foreach(var p in tags ) {
+				if( p.name == revision ) {
+					return p.commit.sha;
+				}
+			}
+			return string.Empty;
+		}
+	}
 }